Hooked more APIs to admin panel. 5 sections left. Attempting to make delete button DRY.

+++ app/Http/Controllers/NewsController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Str;
 use Intervention\Image\Facades\Image;
 Use Illuminate\Support\Facades\File;
+use App\Rules\noimage;
 
 class NewsController extends Controller
 {
@@ -28,11 +29,15 @@ class NewsController extends Controller
         $request->validate([
             'news_title' => ['required', 'string', 'max:255'],
             'news_text' => ['required', 'string'],
-            'news_image' => [ 'image', 'sometimes', 'max:2000'],
-            'resize_image' => ['required', 'numeric', 'integer'],
         ]);
 
         if($request->hasFile('news_image')) {
+
+            $request->validate([
+                'news_image' => [ 'image', 'sometimes', 'max:2000', new noimage],
+                'resize_image' => ['required', 'numeric', 'integer'],
+            ]);
+
             //get filename with extension
             $filenameWithExt = $request->file('news_image')->getClientOriginalName();
             //get just file name (using standard php function)
@@ -81,7 +86,7 @@ class NewsController extends Controller
             return response()->json($news, 200);
         }
 
-        return response()->json('The Provided ID doesn\'t match any News Records !!', 404);
+        return response()->json('The Provided ID Doesn\'t Match Any News Records !!', 404);
     }
 
     public function update(Request $request, $id)
@@ -89,67 +94,73 @@ class NewsController extends Controller
         $request->validate([
             'news_title' => ['required', 'string', 'max:255'],
             'news_text' => ['required', 'string'],
-            'news_image' => [ 'image', 'sometimes', 'max:2000'],
-            'resize_image' => ['required', 'numeric', 'integer'],
         ]);
-
-        if($request->hasFile('news_image')) {
-            //get filename with extension
-            $filenameWithExt = $request->file('news_image')->getClientOriginalName();
-            //get just file name (using standard php function)
-            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
-            //get just extension
-            $extension = $request->file('news_image')->getClientOriginalExtension();
-            //filename to store(uses a time php function to get current time)
-            //this string is a unique name so that file with duplicate name do not get uploaded and
-            //cause problems when viewing(same problem that occured in CISV photo gallery)
-            $filenameToStore= $filename.'_'.time().'.'.Str::lower($extension);
-            
-            // Make Folder if it doesn't exist
-            $path = public_path('storage/news_images');
-            if(!File::isDirectory($path)){
-                    File::makeDirectory($path, 0777, true, true);
-            }
-
-            // Resize image if needed and store it in $image variable
-            // Save image to designated folder inside storage
-            if($request->resize_image == 1){
-                $image = Image::make($request->file('news_image'))->resize(2000, 1000);
-            }
-            else{
-                $image = Image::make($request->file('news_image'));
-            }
-
-            $image->save(public_path('storage/news_images/'. $filenameToStore));
-        }
-        else{
-            $filenameToStore = 'noimage.jpg';
-        }
 
         $news = News::find($id);
         if($news){
-            $news->news_title = $request->news_title;
-            $news->news_text = $request->news_text;            
-            if($request->hasFile('news_image')){  //works if there is a new image uploaded
-                File::delete(public_path('storage/news_images/'.$news->news_image));  //deletes previous file
+                
+            if($request->hasFile('news_image')) {
+
+                $request->validate([
+                    'news_image' => ['required', 'image', 'max:2000', new noimage],
+                    'resize_image' => ['required', 'numeric', 'integer'],
+                ]);
+
+                if($news->news_image != "noimage.jpg")
+                    File::delete(public_path('storage/news_images/'.$news->news_image));  //deletes previous file
+
+                //get filename with extension
+                $filenameWithExt = $request->file('news_image')->getClientOriginalName();
+                //get just file name (using standard php function)
+                $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+                //get just extension
+                $extension = $request->file('news_image')->getClientOriginalExtension();
+                //filename to store(uses a time php function to get current time)
+                //this string is a unique name so that file with duplicate name do not get uploaded and
+                //cause problems when viewing(same problem that occured in CISV photo gallery)
+                $filenameToStore= $filename.'_'.time().'.'.Str::lower($extension);
+                
+                // Make Folder if it doesn't exist
+                $path = public_path('storage/news_images');
+                if(!File::isDirectory($path)){
+                        File::makeDirectory($path, 0777, true, true);
+                }
+
+                // Resize image if needed and store it in $image variable
+                // Save image to designated folder inside storage
+                if($request->resize_image == 1){
+                    $image = Image::make($request->file('news_image'))->resize(2000, 1000);
+                }
+                else{
+                    $image = Image::make($request->file('news_image'));
+                }
+
+                $image->save(public_path('storage/news_images/'. $filenameToStore));
             }
+            else{
+                $filenameToStore = 'noimage.jpg';
+            }
+
+            $news->news_title = $request->news_title;
+            $news->news_text = $request->news_text;
             $news->news_image = $filenameToStore;
             $news->save();
             return response()->json( "News Updated Successfully !!" ,201);
         }
 
-        return response()->json('The Provided ID doesn\'t match any News Records !!', 404);
+        return response()->json('The Provided ID Doesn\'t Match Any News Records !!', 404);
     }
 
     public function destroy($id)
     {
         $news = News::find($id);
         if($news){
-            File::delete(public_path('storage/news_images/'.$news->news_image));  //deletes file
+            if($news->news_image != "noimage.jpg")
+                File::delete(public_path('storage/news_images/'.$news->news_image));  //deletes file
             $news->delete();
             return response()->json( "News Deleted Successfully !!" ,201);
         }
 
-        return response()->json('The Provided ID doesn\'t match any News Records !!', 404);
+        return response()->json('The Provided ID Doesn\'t Match Any News Records !!', 404);
     }
 }
